separate out core vs vrs models
